% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{biplot.gmd}
\alias{biplot.gmd}
\title{The GMD-biplot}
\usage{
\method{biplot}{gmd}(x, ...)
}
\arguments{
\item{x}{An object of class "gmd" that is the output from \code{\link{GMD}}.}

\item{...}{Optional arguments (see Details).}
}
\description{
Biplots based on generelized matrix decomposition.
}
\details{
The optional arguments that can be passed in the biplot are
\itemize{

 \item{\code{index}}{ a numeric vector specifying which variables(arrows) to display. The default is to plot the three longest arrows.}
 \item{\code{names}}{ a string vector specifying the names of the arrows. The default name of the i-th variable is Vi.}
 \item{\code{sample.col}}{ The color of the sample dots. The default is grey.}
 \item{\code{sample.pch}}{ Either an integer specifying a symbol or a single character to be used in plotting sample points. The default is 19.}
 \item{\code{arrow.col}}{ The color of the arrows. The default is orange.}
 \item{\code{arrow.cex}}{ A numeric value giving the size of the labels of the arrows. The default is 1.}

}
}
\examples{
\dontrun{
   X = matrix(rnorm(1000), 50, 20)
   autocorr.mat <- function(p, rho) {
     mat <- diag(p)
     return(rho^abs(row(mat)-col(mat)))
     }
   H = autocorr.mat(50, 0.6)
   Q = autocorr.mat(20, 0.7)
   GMD.fit = GMD(X, H, Q, 2)
   screeplot(GMD.fit)
   biplot(GMD.fit)
   }
}
\references{
Yue Wang, Timothy Randolph, Ali Shojaie and Jing Ma (2019). The GMD-biplot and its application to microbiome data.
}
\seealso{
\code{\link{GMD}}, \code{\link{screeplot.gmd}}
}
\author{
Parker Knight and Yue Wang \email{ywang2310@fredhutch.org}
}
