% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMD.R
\name{fastGMD}
\alias{fastGMD}
\title{Fast GMD Implementation}
\usage{
fastGMD(X, H, Q, K, scale = TRUE)
}
\arguments{
\item{X}{An n x p data matrix.}

\item{H}{An n x n positive semi-definite similarity kernel.}

\item{Q}{An p x p positive semi-definite similarity kernel.}

\item{K}{a scalar specifying the dimension of GMD components (see Details).}

\item{scale}{Logical, indictes whether to scale the data.}
}
\value{
A list of the GMD components U, S, V, H and Q (see Details).
}
\description{
Computes the generalized matrix decomposition of X.
}
\details{
The K-dimensional GMD of X with respect to H and Q is given by X = USV^T, where
\deqn{(U, S, V) = argmin_{U,S,V} ||X - USV^T||^2_{H, Q},}
subject to \eqn{U^THU = I_K, V^TQV = I_K} and \eqn{diag(S) \ge 0}. Here, for any n x p matrix A,
||A||^2_{H,Q} = tr(A^THAQ).
}
\references{
Allen, G. I., L. Grosenick, and J. Taylor (2014). A generalized least-square matrix decom- position. Journal of the American Statistical Association 109(505), 145–159.
}
\author{
Parker Knight and Yue Wang \email{ywang2310@fredhutch.org}
}
