% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMD.R
\name{GMD}
\alias{GMD}
\title{Generalized Matrix Decomposition}
\usage{
GMD(X, H, Q, K)
}
\arguments{
\item{X}{An n x p data matrix.}

\item{H}{An n x n positive semi-definite similarity kernel.}

\item{Q}{An p x p positive semi-definite similarity kernel.}

\item{K}{a scalar specifying the dimension of GMD components (see Details).}
}
\value{
A list of the GMD components U, S, V, H and Q (see Details).
}
\description{
Computes the generalized matrix decomposition of X.
}
\details{
The K-dimensional GMD of X with respect to H and Q is given by X = USV^T, where
\deqn{(U, S, V) = argmin_{U,S,V} ||X - USV^T||^2_{H, Q},}
subject to \eqn{U^THU = I_K, V^TQV = I_K} and \eqn{diag(S) \ge 0}. Here, for any n x p matrix A,
||A||^2_{H,Q} = tr(A^THAQ).
}
\examples{
\dontrun{
   X = matrix(rnorm(1000), 50, 20)
   autocorr.mat <- function(p, rho) {
     mat <- diag(p)
     return(rho^abs(row(mat)-col(mat)))
     }
   H = autocorr.mat(50, 0.6)
   Q = autocorr.mat(20, 0.7)
   GMD.fit = GMD(X, H, Q, 2)
   }

}
\references{
Allen, G. I., L. Grosenick, and J. Taylor (2014). A generalized least-square matrix decom- position. Journal of the American Statistical Association 109(505), 145–159.
}
\author{
Parker Knight and Yue Wang \email{ywang2310@fredhutch.org}
}
