% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{screeplot.gmd}
\alias{screeplot.gmd}
\title{Screeplot of the Generalized Matrix Decomposition}
\usage{
\method{screeplot}{gmd}(fit)
}
\arguments{
\item{fit}{An object of class "gmd" that is the output from \code{\link{GMD}}.}
}
\description{
Screeplot of the Generalized Matrix Decomposition
}
\details{
The function screeplot.gmd plots the proportion of the variances explained by each GMD component of an object of class "gmd".
}
\examples{
\dontrun{
   X = matrix(rnorm(1000), 50, 20)
   autocorr.mat <- function(p, rho) {
     mat <- diag(p)
     return(rho^abs(row(mat)-col(mat)))
     }
   H = autocorr.mat(50, 0.6)
   Q = autocorr.mat(20, 0.7)
   GMD.fit = GMD(X, H, Q, 2)
   screeplot(GMD.fit)
   }
}
\references{
Allen, G. I., L. Grosenick, and J. Taylor (2014). A generalized least-square matrix decom- position. Journal of the American Statistical Association 109(505), 145–159.
}
\seealso{
\code{\link{GMD}}
}
\author{
Parker Knight and Yue Wang \email{ywang2310@fredhutch.org}
}
